% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-address_regex.R
\docType{data}
\name{address_regex}
\alias{address_regex}
\title{Regular expression patterns for address cleaning}
\format{
A dataframe.
}
\source{
Created for use in address cleaning functions.
}
\usage{
address_regex
}
\description{
A table of patterns and other data to use in \code{\link[=clean_address]{clean_address()}}.
}
\keyword{datasets}
