% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_bll_class.R
\name{assign_bll_class}
\alias{assign_bll_class}
\title{Assign BLL classes to lead tests}
\usage{
assign_bll_class(df)
}
\arguments{
\item{df}{A dataframe of lead test records.}
}
\value{
A dataframe with the new variable \code{bll_class}.
}
\description{
This function assigns each lead test to a BLL class based on the KDHE Elevated Blood Lead Disease Investigation Guidelines.
}
