% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_confirmed_status.R
\name{assign_confirmed_status}
\alias{assign_confirmed_status}
\title{Assign confirmed status to lead tests}
\usage{
assign_confirmed_status(
  df,
  row_id,
  rule = c("cdc", "tracking"),
  start_date,
  silent = FALSE
)
}
\arguments{
\item{df}{A dataframe of cleaned, deduplicated lead test records.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{rule}{A string indicating which rule to follow in assigning confirmed status to capillary tests. A capillary test is confirmed if it occurs within \emph{n} days of a prior capillary test.
\itemize{
\item \code{cdc}: \emph{n} = 84
\item \code{tracking}: \emph{n} = 90
}}

\item{start_date}{A string formatted "YYYY-MM-DD" representing the start of the date range covered by the data set.}

\item{silent}{Logical: silence progress indicator if \code{TRUE}.}
}
\value{
The input dataframe with the new variables \code{test_gap_prev} (in days since the previous test), either \code{confirmed_status_cdc} or \code{confirmed_status_tracking}, and \code{test_reason}.
}
\description{
This function subsets \code{df} into test sequences for each person and returns the following variables:
\itemize{
\item the confirmed status (either \code{confirmed_status_cdc} or \code{confirmed_status_tracking})
\item \code{test_gap_prev} (the number of days since the previous test in the sequence)
\item \code{test_gap_next} (the number of days before the next test in the sequence)
\item \code{test_reason}
}
}
\details{
Within the function, tests are subset into sequences for each person in \code{df} and sorted by 1) collection date, 2) specimen source (capillary tests before venous tests), and 3) result value (in descending order).

Confirmed status values:
\itemize{
\item \code{confirmed}: any venous test, or a capillary test following a prior capillary test within \emph{n} days (\emph{n} = 84 for the CDC; \emph{n} = 90 for Tracking)
\item \code{unconfirmed}: a capillary test that has no prior test within \emph{n} days
\item \code{cap_after_ven}: a capillary test following a venous test within \emph{n} days
\item \code{unknown_early}: a capillary test that has no prior test and occurs within \emph{n} days of the start of the data set’s date range (therefore, a prior test may exist outside of the date range of the data set)
\item \code{unknown}: any test not assigned an above status
}

\code{test_reason} values:
\itemize{
\item \code{screening}: a capillary test with no prior test within one year
\item \code{confirmatory}: the next test (venous or capillary) after an elevated screening test
}
}
