% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_elevated_status.R
\name{assign_elevated_status}
\alias{assign_elevated_status}
\title{Assign elevated status to lead tests}
\usage{
assign_elevated_status(df)
}
\arguments{
\item{df}{A dataframe of cleaned, deduplicated lead test records.}
}
\value{
A dataframe with the new variables \code{blrv} and \code{elevated_status}.
}
\description{
This function determines whether or not a test result is elevated according to the blood lead reference value (BLRV) at the time of the blood sample collection.
}
\details{
The BLRV is 3.5 μg/dL for any test collected on or after 2022-01-01. Prior to that, the BLRV was 5 μg/dL.
}
