% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_md_url.R
\name{build_md_url}
\alias{build_md_url}
\title{Build a Melissa Data Personator REST request URL}
\usage{
build_md_url(df, row_id)
}
\arguments{
\item{df}{A dataframe of cleaned addresses.}

\item{row_id}{A unique row identifier variable in \code{df}.}
}
\value{
A vector of URLs.
}
\description{
This function builds a URL to query the Melissa Data Personator Consumer Web Service API. It assumes that the dataframe contains variables named \code{street}, \code{unit}, \code{city}, \code{state}, and \code{zip} which contain the corresponding address components.
}
\section{Address validation workflow}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service.
\enumerate{
\item Prepare addresses
\itemize{
\item \code{\link[=pull_addresses]{pull_addresses()}}
}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_address]{clean_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item Validate/geocode addresses
\itemize{
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=send_md_request]{send_md_request()}}
}
\item Review results
\itemize{
\item \code{\link[=md_results_table]{md_results_table()}}
}
}
}

\section{Melissa Data}{

To check addresses, one of the follow combinations of address fields must be provided at a minimum:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

For more information on constructing the request, see the \href{https://www.melissa.com/quickstart-guides/personator-consumer}{Personator Consumer Quick Start Guide} and the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer:Request}{Personator Consumer Request wiki page}. Additional information can be found at the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer}{Melissa Data wiki}.
}

\seealso{
Other address processing functions: 
\code{\link{clean_address}()},
\code{\link{md_results_table}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{send_md_request}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
