% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age}
\usage{
calculate_age(date1, date2)
}
\arguments{
\item{date1}{A vector containing the birth or start date for the age calculation.}

\item{date2}{A vector containing the end date for the age calculation.}
}
\value{
A numeric vector.
}
\description{
Calculate age from two date vectors. The formula used is \code{(date2 - date1) / 365.25}.
}
\examples{
start <- sample(
  seq.Date(
    from = as.Date("1970-01-01"),
    to = as.Date("1980-01-01"),
    by = "day"
  ),
  size = 10
)

end <- sample(
  seq.Date(
    from = as.Date("1990-01-01"),
    to = as.Date("2020-01-01"),
    by = "day"
  ),
  size = 10
)

calculate_age(date1 = start, date2 = end)

}
