% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_address_table.R
\name{cbls_address_table}
\alias{cbls_address_table}
\title{Create a CBLS Address table}
\usage{
cbls_address_table(df, key, registry)
}
\arguments{
\item{df}{A dataframe of records prepared for CBLS submission.}

\item{key}{A dataframe returned by \code{\link[=cbls_table_key]{cbls_table_key()}}.}

\item{registry}{The CBLS Address Registry.}
}
\value{
A dataframe formatted as an Address table per CBLS guidelines.
}
\description{
Create a CBLS Address table
}
