% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_child_table.R
\name{cbls_child_table}
\alias{cbls_child_table}
\title{Create a CBLS Child table}
\usage{
cbls_child_table(df, key, row_id)
}
\arguments{
\item{df}{A dataframe of records prepared for CBLS submission.}

\item{key}{A dataframe returned by \code{\link[=cbls_table_key]{cbls_table_key()}}.}

\item{row_id}{A unique row identifier variable in \code{df}.}
}
\value{
A dataframe formatted as a Child table per CBLS guidelines.
}
\description{
Create a CBLS Child table
}
