% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_investigation_table.R
\name{cbls_investigation_table}
\alias{cbls_investigation_table}
\title{Create a CBLS Investigation table}
\usage{
cbls_investigation_table(df, key)
}
\arguments{
\item{df}{A dataframe of investigation records.}

\item{key}{A dataframe returned by \code{\link[=cbls_table_key]{cbls_table_key()}}.}
}
\value{
A dataframe formatted as an Investigation table per CBLS guidelines.
}
\description{
Create a CBLS Investigation table
}
