% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_lab_table.R
\name{cbls_lab_table}
\alias{cbls_lab_table}
\title{Create a CBLS Lab Results table}
\usage{
cbls_lab_table(df, row_id, key, ref_lab_type, ref_scrn_site)
}
\arguments{
\item{df}{A dataframe of records prepared for CBLS submission.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{key}{A dataframe returned by \code{\link[=cbls_table_key]{cbls_table_key()}}.}

\item{ref_lab_type}{A reference list of lab names and their classification numbers for the LAB_TYPE column in the Lab Results table.}

\item{ref_scrn_site}{A reference list of facility names and their classification numbers for the SCRN_SITE column in the Lab Results table.}
}
\value{
A dataframe formatted as a Lab Results table per CBLS guidelines.
}
\description{
Create a CBLS Lab Results table
}
