% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_registry_merge.R
\name{cbls_registry_merge}
\alias{cbls_registry_merge}
\title{Merge new CBLS data with the relevant registry}
\usage{
cbls_registry_merge(df_data, df_reg, type = c("address", "child"))
}
\arguments{
\item{df_data}{A dataframe of addresses or children produced during the CBLS data processing workflow.}

\item{df_reg}{A dataframe representing the most recently updated CBLS address or child registry.}

\item{type}{The type of registry being updated. One of \code{c("address", "child")}.}
}
\value{
A list containing
\itemize{
\item \code{df_data}: \code{df_data} plus the registry ID variable
\item \code{df_registry}: \code{df_reg} plus the new addresses or children from \code{df_data}
}
}
\description{
This function merges data (either geocoded addresses or children) produced by the CBLS data processing workflow with the relevant CBLS registry. Registry IDs are assigned to items already present in the registry, and new IDs are created for items new to the registry.
}
