% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_table_key.R
\name{cbls_table_key}
\alias{cbls_table_key}
\title{Populate key variables for CBLS tables}
\usage{
cbls_table_key(df, action)
}
\arguments{
\item{df}{A dataframe of records ready for submission to CBLS.}

\item{action}{The action for each record in the formatted table: \code{"add"}, \code{"change"}, or \code{"delete"}.}
}
\value{
A dataframe with one row and four variables: \code{ACTION}, \code{QTR}, \code{RPT_YR}, and \code{PGMID}.
}
\description{
This function populates values for the second through fifth variables (\code{ACTION} through \code{PGMID}) in each of the CBLS tables. A dataframe of one row is returned, which is then provided in the \code{key} argument of each CBLS table formatting function. Because the submission is quarterly, the function expects data from only one quarter of one year to be provided in \code{df}.
}
\examples{
df <- data.frame(
  lab_collection_date = sample(
    seq.Date(
      from = as.Date("2023-01-01"),
      to = as.Date("2023-03-31"),
      by = "day",
    ),
    size = 10
  )
)
cbls_key <- cbls_table_key(df, action = "add")

}
