% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_write_fwf.R
\name{cbls_write_fwf}
\alias{cbls_write_fwf}
\title{Create a fixed-width file for the CBLS submission}
\usage{
cbls_write_fwf(df, path)
}
\arguments{
\item{df}{A dataframe returned by one of the CBLS table functions and checked by \code{\link[=cbls_check_table]{cbls_check_table()}}.}

\item{path}{A string consisting of a file path and file name. Must end with \code{".txt"}.}
}
\value{
A fixed-width file written by \code{\link[gdata:write.fwf]{gdata::write.fwf()}} exported to the path specified.
}
\description{
Create a fixed-width file for the CBLS submission
}
