% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_seq.R
\name{check_date_seq}
\alias{check_date_seq}
\title{Check date sequences}
\usage{
check_date_seq(date1, date2, ...)
}
\arguments{
\item{date1, date2}{Vectors containing dates. Elements at the same index will be compared, i.e., \code{date1[1]} and \code{date2[1]}.}

\item{...}{Additional date vectors to compare.}
}
\value{
A logical vector of the same length as the arguments.
}
\description{
This function checks whether or not a sequence of dates is in order from earliest to latest. Any number of date vectors can be checked. The order in which arguments are evaluated is \code{date1}, \code{date2}, and then any additional arguments from left to right. Equivalent dates are evaluated as being in sequence. Sequences with \code{NA}s are evaluated with \code{NA}s removed.
}
\examples{
dates <- seq.Date(
  as.Date("2023-01-01", "\%Y-\%m-\%d"),
  as.Date("2023-12-31", "\%Y-\%m-\%d"),
  by = "day"
)
dates <- c(dates, rep(NA, times = 10))

n_rows <- 500

df <- data.frame(
  d1 = sample(dates, n_rows, replace = TRUE),
  d2 = sample(dates, n_rows, replace = TRUE),
  d3 = sample(dates, n_rows, replace = TRUE),
  d4 = sample(dates, n_rows, replace = TRUE)
)

df$ordered <- check_date_seq(df$d1, df$d2, df$d3, df$d4)

}
