% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pattern_df.R
\name{check_pattern_df}
\alias{check_pattern_df}
\title{Search a dataframe for a pattern}
\usage{
check_pattern_df(df, pattern)
}
\arguments{
\item{df}{A dataframe.}

\item{pattern}{A search pattern consisting of regular expressions.}
}
\value{
A message indicating which columns, if any, contain the pattern.
}
\description{
Search each column of a dataframe for a pattern using regular expressions.
}
\examples{
df <- data.frame(
  x = c("one", "two", "three"),
  y = c("eleven", "twelve", "thirteen"),
  z = c("twenty-one", "twenty-two", "twenty-three")
)

check_pattern_df(df, "^one$")

}
