% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_test_reason.R
\name{classify_test_reason}
\alias{classify_test_reason}
\title{Title}
\usage{
classify_test_reason(df1, df2 = NULL, bl_ref_val, max_interval)
}
\arguments{
\item{df1}{The dataframe targeted for classification.}

\item{df2}{A dataframe classified previously.}

\item{max_interval}{}
}
\description{
Title
}
