% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean dirty addresses}
\usage{
clean_address(df, type, var = "street", row_id = "address_id")
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{type}{The type of cleaning to perform. One of \code{c("digit_char", "dir_street", "embed_punct", "house_frac", "nondigit", "nonres", "nth", "num_dir", "num_only", "num_word", "numsign", "phone", "pobox", "rep_seg", "symbol", "unit", "unknown")}.}

\item{var}{A variable in \code{df} containing the targeted address component.}

\item{row_id}{A unique row identifier variable in \code{df}. Defaults to \code{address_id}.}
}
\value{
A dataframe containing values that match the selected error pattern and the suggested replacements.
}
\description{
This function cleans addresses by using regular expressions to match common error patterns. Each \code{type} option corresponds to a pattern in \code{address_regex} that targets text for removal and provides a suggestion for replacement. The function returns a dataframe with two new variables, \code{removed_text} and \code{replacement_text}, adjacent to the variable targeted for cleaning. Only rows containing values that match the error pattern are returned. The user should examine the returned dataframe to confirm, reject, or modify the suggested replacements and then use \code{\link[=replace_values]{replace_values()}} to apply the corrections to the original dataframe.
}
\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
