% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_values.R
\name{clean_values}
\alias{clean_values}
\title{Clean addresses in a dataframe}
\usage{
clean_values(df, var, id_var, ref)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} containing values to be cleaned.}

\item{id_var}{A variable name in \code{df} that serves as a unique row identifier.}

\item{ref}{A reference file with the variables \code{pattern} and \code{replacement}.}
}
\value{
A dataframe containing only the rows that have been cleaned.
}
\description{
Clean address values based on a reference file containing a \code{pattern} variable and a \code{replacement} variable. The returned dataframe has two new columns, \code{removed_text} and \code{replacement_text}. Use \code{\link[=replace_values]{replace_values()}} to substitute the string in \code{replacement_text} for the original string in \code{var}.
}
\seealso{
Other address processing functions: 
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
