% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_elements.R
\name{compare_elements}
\alias{compare_elements}
\alias{compare_val}
\alias{compare_diff}
\alias{compare_tf}
\alias{compare_fn}
\alias{compare_vectors}
\alias{compare_dfs}
\title{Compare values, vectors, and dataframes}
\usage{
compare_val(x, y, ignore = NULL)

compare_diff(x, y, ignore = NULL)

compare_tf(x, y, ignore = NULL)

compare_fn(x, y, fn, ignore = NULL)

compare_vectors(v1, v2, ignore = NULL, numeric = "values")

compare_dfs(df1, df2, ignore = NULL, numeric = "values")
}
\arguments{
\item{x, y}{Values to be compared.}

\item{ignore}{A vector of values to ignore in the comparison.}

\item{fn}{A function that determines the output of \code{compare_fn()}.}

\item{v1, v2}{Vectors of the same length to be compared.}

\item{numeric}{The type of output when comparing unequal numeric values.
\itemize{
\item \code{"values"} returns both input values in a string separated by "|"
\item \code{"diff"} returns the difference between the input values
}}

\item{df1, df2}{Dataframes of the same dimensions to be compared.}
}
\value{
An object with the same dimensions as the input objects. All functions return \code{NA} when one of the input values is included in the \code{ignore} argument. Otherwise, \code{compare_val()} returns a string containing both input values separated by "|". \code{compare_diff()} returns the difference between numeric inputs. \code{compare_tf()} returns \code{TRUE} if the inputs are identical and \code{FALSE} if the inputs are not. \code{compare_fn()} returns a value depending on the function provided to the argument \code{fn}.
}
\description{
Compare values, vectors, and dataframes
}
\examples{
df1 <- data.frame(
  col1 = letters[1:9],
  col2 = TRUE,
  col3 = as.numeric(1:9)
)

df2 <- data.frame(
  col_1 = c(letters[1:3], LETTERS[4:6], letters[7:9]),
  col_2 = NA,
  col_3 = as.numeric(c(1:2, 30, 4:6, 70, 80, 9))
)

comp1 <- compare_dfs(df1, df2)
comp2 <- compare_dfs(df1, df2, ignore = NA)
comp3 <- compare_dfs(df1, df2, ignore = NA, numeric = "diff")

fn_new <- function(x, y) {
  paste("x is", x, "but y is", y)
}

compare_fn(1, 2, fn_new)

}
