% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_parsed_street.R
\name{compare_parsed_street}
\alias{compare_parsed_street}
\title{Compare the street address pre- and post-parsing}
\usage{
compare_parsed_street(
  df,
  street1 = "street_unit",
  street2 = "street_unit_final"
)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{street1}{The pre-parsing street variable.}

\item{street2}{The post-parsing street variable.}
}
\value{
A dataframe containing rows in which \code{street1 != street2}.
}
\description{
Compare the street address pre- and post-parsing
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item \code{\link[=compare_parsed_street]{compare_parsed_street()}}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=submit_to_md]{submit_to_md()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{submit_to_md}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
