% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_epitrax.R
\name{config_epitrax}
\alias{config_epitrax}
\title{Configure the raw EpiTrax data file}
\usage{
config_epitrax(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A list containing two dataframes: \code{data} and \code{keys}.
}
\description{
Adds the variables \code{recno}, a unique identifier for each row, and \code{hash_value}, a unique identifier for the combination of variables in each row. Formats all date columns.
}
