% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_conflicts.R
\name{count_conflicts}
\alias{count_conflicts}
\title{Count dupesets with conflicting values}
\usage{
count_conflicts(df, silent = FALSE)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}

\item{silent}{Logical: silence output to console if \code{TRUE}.}
}
\value{
A dataframe with one row per variable in \code{df}. The returned dataframe has three columns:
\itemize{
\item \code{var_name}: The variable name from \code{df}.
\item \code{conflict_count}: The count of dupesets in \code{df} with conflicting values.
\item \code{conflict_percent}: The percentage of dupesets in \code{df} with conflicting values.
}
}
\description{
For each variable in a dataframe, \code{count_conflicts()} counts the number of dupesets with any conflicting values.
}
\section{Deduplication}{

Deduplication is the process of reducing a dataframe to distinct rows only. Rows that have identical values across a set of selected variables are considered duplicates. These selected variables are "visible" to the process of deduplication. All other variables in the dataframe are "invisible" to deduplication, so they are ignored. Within dupesets, variables that were invisible to deduplication can have conflicting values.

The \code{\link[=undupe]{undupe()}} function produces both a deduplicated dataframe and a dataframe of all duplicate rows (dupesets) pulled from the original dataframe.

For a more detailed explanation of deduplication, see the \code{\link[=undupe]{undupe()}} documentation.
}

\examples{
n_rows <- 20
df <- data.frame(
  x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
  y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
  z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE)
)
undupe <- undupe(df, visible_vars = c("x", "y"))
df_count <- count_conflicts(undupe[["df_dupesets"]])

}
\seealso{
Other undupe functions: 
\code{\link{dupes2xl}()},
\code{\link{flatten_conflicts}()},
\code{\link{isolate_conflicts}()},
\code{\link{undupe2}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
