% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_dupeset_diffs.R
\name{count_dupeset_diffs}
\alias{count_dupeset_diffs}
\title{Count dupesets with value differences for each variable}
\usage{
count_dupeset_diffs(df)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}
}
\value{
A table with one row per variable in \code{df} and both the count and percentage of dupesets in which value differences occur for each variable.
}
\description{
When a dataframe is deduplicated, values may differ among duplicates in the variables that were not used as the basis for deduplication. \code{count_dupeset_diffs()} tallies the number of dupesets in which value differences occur for each variable in \code{df}.
}
\examples{
n_rows <- 20
df <- data.frame(x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
                 y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
                 z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE))
undupe <- undupe(df, undupe_vars = c("x", "y"))
df_diff_count <- count_dupeset_diffs(undupe[["df_dupesets"]])
}
\seealso{
Other undupe functions: 
\code{\link{aggregate_values}()},
\code{\link{distill_dupeset_diffs}()},
\code{\link{dupes2xl}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
