% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bl_data_project.R
\name{create_bl_data_project}
\alias{create_bl_data_project}
\title{Create blood lead project and subdirectories}
\usage{
create_bl_data_project(parent_dir, data_range)
}
\arguments{
\item{parent_dir}{The path to the directory in which the BL project will reside.}

\item{data_range}{The date range of the BL data that will be processed.
\itemize{
\item For one quarter, use the format YYYYqQ (e.g., "2023q1")
\item For one year, use the format YYYY (e.g., "2023")
\item For a range that spans multiple years, use the format YYYY-YYYY (e.g., "2015-2022")
}}
}
\value{
A message indicating success or failure in creating all directories.
}
\description{
This function creates an R project and subdirectories for a blood lead data processing project. If \code{parent_dir} is not found or if a project directory with the same name already exists, the function will error out before creating anything. The R project is created using \code{\link[usethis:create_package]{usethis::create_project()}}.
}
\examples{
\dontrun{
create_bl_data_project(
  parent_dir = "~/my_projects/",
  data_range = "2023q1"
)
}

}
