% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_dirs.R
\name{create_project_dirs}
\alias{create_project_dirs}
\title{Create project directories}
\usage{
create_project_dirs(type = c("analysis1", "analysis2", "ksde"), root = getwd())
}
\arguments{
\item{type}{The type of project structure to create. One of \code{c("analysis1", "analysis2", "ksde")}. All create three subdirectories: "data", "data/1-source", "output", and "scripts". The differences are that
\itemize{
\item \code{"analysis1"} also creates "data/2-aux" and "data/3-final".
\item \code{"analysis2"} also creates "data/2-final".
\item \code{"ksde"} also creates "data/2-aux", "data/3-final", "output/1-test", and "output/2-prod".
}}

\item{root}{The project root directory. Defaults to \code{getwd()}.}
}
\value{
Returns the project root directory, invisibly.
}
\description{
Create project directories
}
