% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a report spreadsheet}
\usage{
create_report(df, file, style_cols, overwrite)
}
\arguments{
\item{df}{A dataframe.}

\item{file}{A file path. \code{".xlsx"} will be appended if not present.}

\item{style_cols}{A vector of variable names in \code{df} to highlight.}

\item{overwrite}{Logical: overwrite existing file if \code{TRUE}.}
}
\value{
An Excel workbook.
}
\description{
This function wraps around \code{\link[=my_xl_table]{my_xl_table()}} to create a report spreadsheet in Excel. It adds a column called "action" and suppresses the error returned by \code{openxlsx::saveWorkbook} when the file path already exists and \code{overwrite == FALSE}.
}
