% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-directions.R
\docType{data}
\name{directions}
\alias{directions}
\title{Street directions}
\format{
A dataframe.
}
\source{
Created for use in address cleaning and address creation functions.
}
\usage{
directions
}
\description{
A table of street directions, with both full and abbreviated forms.
}
\keyword{datasets}
