% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupeset_conflict_focus.R
\name{distill_dupeset_diffs}
\alias{distill_dupeset_diffs}
\title{Distill dupesets to just those with value differences in selected variables}
\usage{
distill_dupeset_diffs(df, vars, silence = FALSE)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}

\item{vars}{A character vector of variable names in \code{df}.}

\item{silence}{Logical: silence progress bar if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
When a dataframe is deduplicated, values may differ among duplicates in the variables that were not used as the basis for deduplication. \code{distill_dupeset_diffs()} produces a dataframe that keeps only those dupesets that have value differences in the variables named in \code{vars}. This makes it easier to visually inspect the data and look for patterns among dupesets, especially when \code{\link[=dupes2xl]{dupes2xl()}} is run on the distilled dataframe.
}
\examples{
n_rows <- 20
df <- data.frame(x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
                 y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
                 z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE))
undupe <- undupe(df, undupe_vars = c("x", "y"))
df_distilled <- distill_dupeset_diffs(undupe[["df_dupesets"]], vars = "z")
}
\seealso{
Other undupe functions: 
\code{\link{aggregate_values}()},
\code{\link{dupes2xl}()},
\code{\link{dupeset_conflict_count}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
