% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_ksde_project.R
\name{document_ksde_project}
\alias{document_ksde_project}
\title{Document a KS Tracking Data Explorer project}
\usage{
document_ksde_project(
  proj_dir,
  proj_type = c("rabies", "geo", "mw", "version"),
  doc_type = c("code", "output"),
  overwrite = FALSE,
  test_drive = FALSE
)
}
\arguments{
\item{proj_dir}{The name of the RStudio project to document.}

\item{proj_type}{The project type.
\itemize{
\item \code{"rabies"} = A Rabies Surveillance display data project.
\item \code{"geo"} = A core geography tables project.
\item \code{"mw"} = A message widget table project.
\item \code{"version"} = A version table project.
}}

\item{doc_type}{The part of the project to document.
\itemize{
\item \code{"code"} = All parts of the project necessary to reproduce the output, including the project folder structure, source data, and scripts.
\item \code{"output"} = The project output only.
}}

\item{overwrite}{If the destination directory for the project already exists, it will be removed when \code{overwrite} is set to \code{TRUE}.}

\item{test_drive}{Replace the destination path with the desktop ("C:/Users/eliot.monaco/OneDrive - State of Kansas, OITS/Desktop/") if \code{TRUE}.}
}
\value{
No value returned.
}
\description{
Document a KS Tracking Data Explorer project
}
