% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes2xl.R
\name{dupes2xl}
\alias{dupes2xl}
\title{Export dupesets to an Excel file for easy viewing}
\usage{
dupes2xl(df, filename, path = NULL, visible = FALSE)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}

\item{filename}{A name for the new Excel workbook.}

\item{path}{The folder path containing the Excel workbook to be used as a template.}

\item{visible}{A logical value: view the workbook while the macros run if \code{TRUE}.}
}
\value{
Nothing is returned in R. An Excel workbook is created using the \code{filename} and \code{path} provided as function arguments.
}
\description{
This function saves a dataframe of dupesets to an existing macro-enabled Excel workbook. This workbook contains VBA code to format the data, making dupesets easy to differentiate and value differences easy to detect.
}
\examples{
n_rows <- 20
df <- data.frame(
  x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
  y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
  z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE)
)
undupe <- undupe(df, visible_vars = c("x", "y"))
f <- ""
p <- ""
\dontrun{
dupes2xl(undupe[["df_dupesets"]], filename = f, path = p)
}
}
\seealso{
Other undupe functions: 
\code{\link{count_conflicts}()},
\code{\link{flatten_conflicts}()},
\code{\link{isolate_conflicts}()},
\code{\link{undupe2}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
