% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupeset_conflict_count.R
\name{dupeset_conflict_count}
\alias{dupeset_conflict_count}
\title{For each dataframe variable count the number of dupesets with conflicting values}
\usage{
dupeset_conflict_count(df)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}
}
\value{
A table with one row per variable in \code{df} and both the count and percentage of dupesets in which value differences occur for each variable.
}
\description{
Dupesets are records that have the same values in selected variables. However, within a dupeset, any other variables may contain conflicting values. \code{dupeset_conflict_count()} tallies the number of dupesets with conflicting values for each variable in \code{df}.
}
\examples{
n_rows <- 20
df <- data.frame(
  x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
  y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
  z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE)
)
undupe <- undupe(df, undupe_vars = c("x", "y"))
df_diff_count <- dupeset_conflict_count(undupe[["df_dupesets"]])
}
\seealso{
Other undupe functions: 
\code{\link{aggregate_values}()},
\code{\link{distill_dupeset_diffs}()},
\code{\link{dupes2xl}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
