% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupeset_conflict_focus.R
\name{dupeset_conflict_focus}
\alias{dupeset_conflict_focus}
\title{Focus on dupesets with conflicting values in selected variables}
\usage{
dupeset_conflict_focus(df, vars, silent = FALSE)
}
\arguments{
\item{df}{A dataframe of dupesets returned by \code{\link[=undupe]{undupe()}}.}

\item{vars}{A character vector of variable names in \code{df}.}

\item{silent}{Logical: silence progress bar if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
Dupesets are records that have the same values in selected variables. However, within a dupeset, any other variables may contain conflicting values. \code{dupeset_conflict_focus()} produces a dataframe that keeps only those dupesets that have conflicting values in the variables named in \code{vars}. This makes it easier to visually inspect the data and look for patterns among dupesets, especially when \code{\link[=dupes2xl]{dupes2xl()}} is run on the returned dataframe.
}
\examples{
n_rows <- 20
df <- data.frame(
  x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
  y = sample(c(1, 10, 100, NA), size = n_rows, replace = TRUE),
  z = sample(c("banana", "carrot", "pickle"), size = n_rows, replace = TRUE)
)
undupe <- undupe(df, undupe_vars = c("x", "y"))
df_distilled <- dupeset_conflict_focus(undupe[["df_dupesets"]], vars = "z")
}
\seealso{
Other undupe functions: 
\code{\link{aggregate_values}()},
\code{\link{dupes2xl}()},
\code{\link{dupeset_conflict_count}()},
\code{\link{undupe}()}
}
\concept{undupe functions}
