% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epitrax_lead_vars.R
\docType{data}
\name{epitrax_lead_vars}
\alias{epitrax_lead_vars}
\title{EpiTrax variables used in lead data processing}
\format{
A dataframe.
}
\source{
Created for lead data processing.
}
\usage{
epitrax_lead_vars
}
\description{
A list of vectors containing variables used in the lead data processing pipeline, including in \code{\link[=config_epitrax]{config_epitrax()}}.
}
\keyword{datasets}
