% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epitrax_vars.R
\docType{data}
\name{epitrax_vars}
\alias{epitrax_vars}
\title{EpiTrax blood lead record variables}
\format{
A list.
}
\source{
EpiTrax export file.
}
\usage{
epitrax_vars
}
\description{
Variable names contained in the EpiTrax lead data export file. Used in \code{\link[=config_epitrax]{config_epitrax()}}. The list contains three vectors:
\itemize{
\item \code{all}: All exported variables
\item \code{date}: Variables to format as date
\item \code{final}: Variables kept in the final data set
}
}
\keyword{datasets}
