% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etmstuff-package.R
\docType{package}
\name{etmstuff-package}
\alias{etmstuff}
\alias{etmstuff-package}
\title{etmstuff: All my custom functions in one place}
\description{
Functions with broad applications and functions for handling lead program data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eliotmonaco/etmstuff}
  \item Report bugs at \url{https://github.com/eliotmonaco/etmstuff/issues}
}

}
\author{
\strong{Maintainer}: Eliot Monaco \email{eliotmonaco@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
