% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_elevated_results.R
\name{flag_elevated_results}
\alias{flag_elevated_results}
\title{Flag lead tests for elevated results}
\usage{
flag_elevated_results(df, blrv)
}
\arguments{
\item{df}{A dataframe of lead test records.}

\item{blrv}{Numeric: the blood lead reference value (mcg/dL) in effect during the date range of the data provided in \code{df}. A blood lead test result >= \code{blrv} is elevated.}
}
\value{
A dataframe with the new variables \code{bl_ref_val} and \code{lab_result_elev}.
}
\description{
This function determines whether or not the test result is elevated according to the blood lead reference value.
}
