% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_overdue_followup.R
\name{flag_overdue_followup}
\alias{flag_overdue_followup}
\title{Flag lead tests for overdue follow-up}
\usage{
flag_overdue_followup(df)
}
\arguments{
\item{df}{A dataframe of lead test records.}
}
\value{
A dataframe with the new variables \code{followup_interval}, \code{sufficient_interval}, and \code{followup_overdue}.
}
\description{
This function determines whether the follow-up for each test is overdue according to the KDHE Elevated Blood Lead Disease Investigation Guidelines. The function is used in the lead dashboard.
}
