% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_compare.R
\name{fuzzy_compare}
\alias{fuzzy_compare}
\title{Compare rows using a combination of fuzzy and exact methods}
\usage{
fuzzy_compare(df1, df2, fuzzy_var, exact_var = NULL, ignore_case = TRUE)
}
\arguments{
\item{df1}{A dataframe with rows to be compared.}

\item{df2}{A dataframe to be compared against.}

\item{fuzzy_var}{Names of variables in \code{df1} and \code{df2} selected for fuzzy comparison.}

\item{exact_var}{Names of variables in \code{df1} and \code{df2} selected for exact matching (optional).}

\item{ignore_case}{Logical. Ignore case when comparing fuzzy and exact variables if \code{TRUE}.}
}
\value{
A dataframe of potential matches joined by the variables in \code{exact_var}, with the new variable \code{sim_score}.
}
\description{
This function compares rows in two dataframes, allowing some values to be matched exactly and other values to be scored based on their similarity. It was designed for comparing addresses, in which certain values such as the city, state, and zip code are more amenable to exact matching and other values such as the street and unit components are subject to a large degree of variation and may require a flexible or inexact method of comparison.

The variables being compared must have the same names in both dataframes. If addresses in \code{df1} are being compared, e.g., \code{df1$street}, \code{df1$city}, and \code{df1$state}, then \code{df2} must have the same variable names, \code{df2$street}, \code{df2$city}, and \code{df2$state}.
}
\examples{
df1 <- sim_address(nrow = 10)
df2 <- sim_address(nrow = 5000)

df_match <- fuzzy_compare(
  df1,
  df2,
  fuzzy_var = c("street", "unit"),
  exact_var = c("city", "state", "zip")
)

}
