% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_cities.R
\docType{data}
\name{ks_cities}
\alias{ks_cities}
\title{A list of city names in Kansas}
\format{
A character vector of Kansas city names.
}
\source{
City names were taken from an initial list in ZipCode_Esri2019.xls. Additional names were added as new cities were found during the address processing phase of the lead data processing pipeline.
}
\usage{
ks_cities
}
\description{
A list of Kansas city names to use in address validation.
}
\keyword{datasets}
