% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_city_zip.R
\docType{data}
\name{ks_city_zip}
\alias{ks_city_zip}
\title{Kansas city, zip code, and county combinations}
\format{
A dataframe.
}
\source{
Downloaded from the Five-digit Zip Code Tax Rate Information web page on the Kansas Department of Revenue site (https://www.ksrevenue.gov/5digitzip.html). Additional rows are added as new combinations are discovered.
}
\usage{
ks_city_zip
}
\description{
A list of associated cities, zip codes, and counties in Kansas.
}
\keyword{datasets}
