% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_county_fips.R
\docType{data}
\name{ks_county_fips}
\alias{ks_county_fips}
\title{Kansas counties and FIPS codes}
\format{
A dataframe with 7 columns and 105 rows.
\describe{
\item{county}{Short county name (does not include "County")}
\item{county_full}{Full county name (includes "County")}
\item{state_fips}{State FIPS code}
\item{county_fips}{County FIPS code}
\item{county_ns}{County NS code}
\item{fips_class}{FIPS class code}
\item{func_stat}{Functional status code}
}
}
\source{
Downloaded from the US Census Bureau website (https://www.census.gov/library/reference/code-lists/ansi.html#cou).
}
\usage{
ks_county_fips
}
\description{
A list of Kansas counties and their corresponding 2020 census codes. Table downloaded from Census.gov.
}
\keyword{datasets}
