% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_fips.R
\docType{data}
\name{ks_fips}
\alias{ks_fips}
\title{State and county FIPS codes in Kansas}
\format{
A dataframe with 6 columns and 105 rows.
\describe{
\item{fips}{5-digit FIPS code (state + county)}
\item{st_fips}{2-digit state FIPS code}
\item{cnty_fips}{3-digit county FIPS code}
\item{cnty_code}{2-letter county abbreviation}
\item{state}{2-letter state abbreviation}
\item{county}{County name}
}
}
\source{
Unknown.
}
\usage{
ks_fips
}
\description{
A list of state and county FIPS codes, county abbreviations, and county names for Kansas.
}
\keyword{datasets}
