% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_place_by_county.R
\docType{data}
\name{ks_place_by_county}
\alias{ks_place_by_county}
\title{Kansas cities and CDPs by county}
\format{
A dataframe with 10 columns and 764 rows.
\describe{
\item{STATE}{State postal abbreviation}
\item{STATEFP}{State FIPS code}
\item{COUNTYFP}{County FIPS code}
\item{COUNTYNAME}{County name and legal/statistical area description}
\item{PLACEFP}{Place FIPS code}
\item{PLACENS}{Place NS code}
\item{PLACENAME}{Place name and legal/statistical area description}
\item{TYPE}{Place type}
\item{CLASSFP}{FIPS class code}
\item{FUNCSTAT}{Legal functional status}
}
}
\source{
https://www.census.gov/library/reference/code-lists/ansi.html#place
}
\usage{
ks_place_by_county
}
\description{
A list of Kansas incorporated and census designated places (CDPs) and their corresponding 2020 census codes. There is at least one row per place. If a place overlaps with multiple counties, there is a separate row for each county. Table downloaded from Census.gov.
}
\keyword{datasets}
