% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ks_zipcodes.R
\docType{data}
\name{ks_zipcodes}
\alias{ks_zipcodes}
\title{A list of zip codes in Kansas}
\format{
A character vector of Kansas zip codes.
}
\source{
Zip codes were taken from an initial list in ZipCode_Esri2019.xls. Additional zip codes were added as new codes were found during the address processing phase of the lead data processing pipeline.
}
\usage{
ks_zipcodes
}
\description{
A list of Kansas zip codes to use in address validation.
}
\keyword{datasets}
