% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_elements.R
\name{match_elements}
\alias{match_elements}
\title{Check for matching elements in two vectors}
\usage{
match_elements(v1, v2)
}
\arguments{
\item{v1}{A vector of any type.}

\item{v2}{A vector of any type.}
}
\value{
Returns a message stating if elements in the vectors match, and if not, returns a list of which elements do not.
}
\description{
Check for matching elements in two vectors
}
\examples{
x <- sample(state.name, size = 40)
y <- sample(state.name, size = 40)
match_elements(x, y)

x <- sample(1:30, size = 20)
y <- sample(1:30, size = 20)
match_elements(x, y)
}
