% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_request.R
\name{md_request}
\alias{md_request}
\alias{md_batch_request}
\title{Send requests to the Melissa Data Personator API}
\usage{
md_batch_request(df, var_url = "md_url")

md_request(url)
}
\arguments{
\item{df}{A dataframe.}

\item{var_url}{The name of the variable in \code{df} containing the URL requests for Melissa Data Personator. Defaults to \code{md_url}.}

\item{url}{A single URL request for Melissa Data Personator.}
}
\value{
A dataframe with a single or multiple responses from Melissa Data Personator.
}
\description{
These functions send the URL requests created by \code{\link[=build_md_url]{build_md_url()}} to the Melissa Data Personator Consumer Web Service API. \code{md_request()} sends a single request. \code{md_batch_request()} sends multiple requests and adds the responses from Melissa Data to \code{df} as additional columns.
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item \code{\link[=compare_parsed_street]{compare_parsed_street()}}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}

Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
