% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_results_table.R
\name{md_results_table}
\alias{md_results_table}
\title{Count result codes in the Melissa Data response}
\usage{
md_results_table(df, var = "Results")
}
\arguments{
\item{df}{A dataframe of Melissa Data responses returned by \code{\link[=md_batch_request]{md_batch_request()}}.}

\item{var}{The variable containing result codes. The default is \code{"Results"}.}
}
\value{
A dataframe tallying addresses by certain Melissa Data result codes.
}
\description{
This function creates a table with counts of the result codes that are most relevant to the lead address validation and geocoding process. The following result codes or result code combinations are included: AS01, AS01 + AS23, AS01 + AS26, AS02, AS13, AE02, AE06, , , , , , ,
}
