% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-melissa_data_codes.R
\docType{data}
\name{melissa_data_codes}
\alias{melissa_data_codes}
\title{Melissa Data result codes}
\format{
A list of five dataframes.
}
\source{
Copied on 2024-02-01 from the \href{https://wiki.melissadata.com/index.php?title=Result_Code_Details#Personator_Consumer}{Melissa Wiki web page}.
}
\usage{
melissa_data_codes
}
\description{
Addresses sent to Melissa Data for validation are returned with result codes in the \code{Results} column. These codes provide information about the validity of the submitted address and indicate if any errors were found or if any changes were made in the response. \code{melissa_data_codes} is a list of five tables:
\itemize{
\item \code{address_status}
\item \code{address_error}
\item \code{address_change}
\item \code{geocode_status}
\item \code{geocode_error}
}
}
\keyword{datasets}
