% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-melissa_data_result_codes.R
\docType{data}
\name{melissa_data_result_codes}
\alias{melissa_data_result_codes}
\title{Melissa Data result codes}
\format{
A list containing five dataframes.
}
\source{
\href{https://wiki.melissadata.com/index.php?title=Result_Code_Details#Personator_Consumer}{Melissa Wiki - Result Code Details}
}
\usage{
melissa_data_result_codes
}
\description{
Addresses sent to Melissa Data for validation are returned with result codes in the \code{Results} column. These codes provide information about the validity of the address and indicate if errors were found or changes were made to the submitted address. \code{melissa_data_result_codes} is a list containing five tables, each of which is a different result code category:
\itemize{
\item \code{address_status}
\item \code{address_error}
\item \code{address_change}
\item \code{geocode_status}
\item \code{geocode_error}
}
}
\keyword{datasets}
