% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_address_registry.R
\name{merge_address_registry}
\alias{merge_address_registry}
\title{Merge new cleaned addresses with the CBLS address registry}
\usage{
merge_address_registry(df, registry)
}
\arguments{
\item{df}{A dataframe of cleaned addresses prepared for the CBLS submission.}

\item{registry}{The most recent CBLS address registry.}
}
\value{
A list consisting of \code{df_addr} (\code{df} plus the variable \code{address_registry_id}), and \code{df_registry} (the address registry plus any new addresses from \code{df}).
}
\description{
Merge new cleaned addresses with the CBLS address registry
}
