% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_child_registry.R
\name{merge_child_registry}
\alias{merge_child_registry}
\title{Merge new records with the CBLS child registry}
\usage{
merge_child_registry(df, registry)
}
\arguments{
\item{df}{A dataframe of records prepared for the CBLS submission.}

\item{registry}{The most recent CBLS child registry.}
}
\value{
A list consisting of \code{df_data} (\code{df} plus the variable \code{child_registry_id}), and \code{df_registry} (the child registry plus any new children from \code{df}).
}
\description{
Merge new records with the CBLS child registry
}
