% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xl_table.R
\name{my_xl_table}
\alias{my_xl_table}
\title{Write data to an Excel workbook}
\usage{
my_xl_table(
  data,
  file,
  style_cols = NULL,
  sheet = "Sheet1",
  as_table = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{data}{A dataframe (or list of dataframes).}

\item{file}{A file path. \code{".xlsx"} will be appended if not present.}

\item{style_cols}{A character vector (or list of vectors) of variable names in \code{data} to highlight.}

\item{sheet}{A name (or vector of names) for the worksheet(s).}

\item{as_table}{Logical: format data as a table if \code{TRUE}.}

\item{overwrite}{Logical: overwrite existing file if \code{TRUE}.}
}
\value{
None.
}
\description{
This function produces an Excel workbook using the \link{openxlsx} package.
}
\details{
If \code{data} is a list of dataframes, each will be written to a separate sheet in a single workbook. Dataframes with 0 rows will be skipped.
}
\examples{
\dontrun{
my_xl_table(mtcars, file = "mtcars_workbook", style_cols = c("mpg", "cyl"))
}

}
