% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xlsx.R
\name{my_xlsx}
\alias{my_xlsx}
\title{Save XLSX file with custom settings using \code{write.xlsx()}}
\usage{
my_xlsx(x, file)
}
\arguments{
\item{x}{A dataframe or a list of dataframes.}

\item{file}{A file path.}
}
\value{
An XLSX file.
}
\description{
Save a dataframe as an XLSX file with table formatting using \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}.
}
