% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-names.R
\docType{data}
\name{names}
\alias{names}
\title{First and last names}
\format{
A list.
}
\source{
names.xlsx
}
\usage{
names
}
\description{
A list of first and last names to use in simulated data. The list contains the following vectors:
\itemize{
\item \code{first_male}: 100 male first names
\item \code{first_female}: 100 female first names
\item \code{last}: 1000 last names
}
}
\keyword{datasets}
