% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_street_addresses.R
\name{parse_street_addresses}
\alias{parse_street_addresses}
\title{Parse street addresses using \code{postmastr} workflow}
\usage{
parse_street_addresses(df, ref)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{ref}{This will be changed...}
}
\value{
TBD...
}
\description{
Parse street addresses using functions from the \code{postmastr} package. \code{df} requires an \code{address_id} variable before using \code{parse_street_addresses()}.

\code{postmastr} is only available in a development version from GitHub.

\if{html}{\out{<div class="sourceCode">}}\preformatted{remotes::install_github("slu-openGIS/postmastr")
}\if{html}{\out{</div>}}
}
\section{Address processing workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item compare...
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=submit_to_md]{submit_to_md()}}
}
}

\section{Address validation and geocoding}{

The address processing workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for the request are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_values}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{submit_to_md}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
