% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_check_df.R
\name{pattern_check_df}
\alias{pattern_check_df}
\title{Check dataframe columns for a pattern}
\usage{
pattern_check_df(df, pattern, ignore_case = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{pattern}{A pattern to search for, formatted as a regular expression.}

\item{ignore_case}{Logical: ignore case in pattern search if \code{TRUE}.}
}
\value{
Returns a message listing which columns contain the pattern.
}
\description{
Search a dataframe for a pattern formatted as a regular expression. For exact matches, use \code{^} and \code{$} at the beginning and ending of the string in \code{pattern}.
}
\examples{
df <- data.frame(
  x = c("one", "two", "three"),
  y = c("eleven", "twelve", "thirteen"),
  z = c("twenty-one", "twenty-two", "twenty-three")
)
pattern_check_df(df, "^one$")

}
