% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Calculate a percentage}
\usage{
pct(n, total, dec = 1)
}
\arguments{
\item{n}{A number. The numerator of the fraction that will become a percentage.}

\item{total}{A number. The denominator of the fraction that will become a percentage.}

\item{dec}{An integer. The number of decimal places to round the percentage to.}
}
\value{
A number.
}
\description{
Calculate a percentage
}
\examples{
pct(5, 25)

x <- sample(200:300, size = 10, replace = TRUE)
pct(x, total = 300, dec = 2)

}
