% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_highest_confirmed_test.R
\name{pick_highest_confirmed_test}
\alias{pick_highest_confirmed_test}
\title{Pick the highest confirmed test for each person in the data set}
\usage{
pick_highest_confirmed_test(df, silent = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{silent}{Logical: silence output to console if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
This function is used when counting individuals based on the blood lead level (BLL) from their test results. It picks the test with the highest BLL, preferring a confirmed over an unconfirmed test, for each person within the time period defined by the data in \code{df}. Tests are grouped by \code{patient_id}. Tests for the same person that occur within 92 days of another belong to one sequence. The highest confirmed test is selected within each person sequence. If a person has more than one sequence, the highest confirmed test among the sequences is then selected.
}
