% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_max_cfm_test.R
\name{pick_max_cfm_test}
\alias{pick_max_cfm_test}
\title{Pick the test (confirmatory preferred) with the highest result for a person}
\usage{
pick_max_cfm_test(df, cfm_only = FALSE, silent = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{cfm_only}{Logical: consider confirmatory tests only if \code{TRUE}; consider all tests (confirmatory, screening, and unknown/other) if \code{FALSE} (the default).}

\item{silent}{Logical: silence progress indicator if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
This function selects one test per person from \code{df}, filtering records by \code{patient_id} and evaluating them based on the values in \code{test_reason} and \code{lab_result_number}. There are two modes:
\enumerate{
\item Consider all tests (\code{cfm_only = FALSE}): After filtering tests by \code{patient_id}, the confirmatory test with the highest result is kept. If there are no confirmatory tests for an individual, the non-confirmatory test with the highest result is kept.
\item Consider confirmatory tests only (\code{cfm_only = TRUE}): After filtering tests by \code{patient_id}, the confirmatory test with the highest result is kept. If there is no confirmatory test for an individual, they are not represented in the output. Non-confirmatory tests are not considered.
}
}
\details{
The \code{test_reason} variable is created by the function \code{\link[=classify_test_reason]{classify_test_reason()}}.

Confirmatory \code{test_reason} values:
\itemize{
\item \code{cap_cfm}
\item \code{ven_cfm_i}
\item \code{ven_cfm_e}
\item \code{ven_cfm_n}
\item \code{ven_flw}
}

Non-confirmatory \code{test_reason} values:
\itemize{
\item \code{cap_scrn}
\item \code{unknown/other}
}
}
