% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_check.R
\name{prox_check}
\alias{prox_check}
\title{Check the proximity of two numbers}
\usage{
prox_check(x, y, pct = 5, n = 2, side = "both")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{pct}{A number defining a range to check as a percent of \code{y}.}

\item{n}{A number defining an absolute range to check.}

\item{side}{The side of \code{y} to check: \code{"above"}, \code{"below"}, or \code{"both"} (default).}
}
\value{
A logical value.
}
\description{
Returns a logical value indicating whether or not \code{x} is within a specified range of \code{y}.
}
\details{
The proximity of \code{x} to \code{y} is checked for two range types: a percent (\code{pct}) of \code{y} and a number (\code{n}). \code{prox_check()} returns \code{TRUE} if \code{x} is within either range. If only one range type should be evaluated, set the other type to \code{0}. \code{side} determines whether the range is evaluated above \code{y}, below \code{y}, or both.
}
\examples{
prox_check(105, 100)

}
