% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_md_request.R
\name{send_md_request}
\alias{send_md_request}
\title{Send requests to the Melissa Data Personator API}
\usage{
send_md_request(url)
}
\arguments{
\item{url}{A character vector containing one or more REST requests for Melissa Data Personator.}
}
\value{
A dataframe.
}
\description{
This function sends URL requests created by \code{\link[=build_md_url]{build_md_url()}} to the Melissa Data Personator Consumer Web Service API.
}
\section{Address validation workflow}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service.
\enumerate{
\item Prepare addresses
\itemize{
\item \code{\link[=pull_addresses]{pull_addresses()}}
}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_address]{clean_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item Validate/geocode addresses
\itemize{
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=send_md_request]{send_md_request()}}
}
\item Review results
\itemize{
\item \code{\link[=md_results_table]{md_results_table()}}
}
}
}

\section{Melissa Data}{

To use the Check action to validate and correct addresses, Melissa Data requires, at minimum, one of the follow combinations of address fields:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

For more information on constructing the request, see the \href{https://www.melissa.com/quickstart-guides/personator-consumer}{Personator Consumer Quick Start Guide} and the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer:Request}{Personator Consumer Request wiki page}. Additional information can be found at the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer}{Melissa Data wiki}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_address}()},
\code{\link{md_results_table}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
