% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_address.R
\name{sim_address}
\alias{sim_address}
\title{Simulate addresses}
\usage{
sim_address(n)
}
\arguments{
\item{n}{The number of addresses in the output.}
}
\value{
A dataframe.
}
\description{
Create a dataframe of simulated Kansas addresses.
}
\details{
Street names are sampled from \code{street_names}. City, zip code, and county combinations are sampled from \code{ks_city_zip}.
}
\examples{
df <- sim_address(n = 1000)

}
