% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lead_data.R
\name{sim_lead_data}
\alias{sim_lead_data}
\title{Simulate lead test records}
\usage{
sim_lead_data(n, dmin, dmax = Sys.Date())
}
\arguments{
\item{n}{The number of people to include in the data set.}

\item{dmin, dmax}{The minimum and maximum of the data set's date range. Must be dates formatted YYYY-MM-DD. \code{dmax} defaults to \code{Sys.Date()}.}
}
\value{
A dataframe.
}
\description{
This function creates a simulated data set of lead test records.
}
\details{
\code{n} determines the number of people in the data set created by \code{sim_person()}, while \code{sim_test_seq()} creates a plausible test sequence consisting of one or more tests for each person within the date range of the data set.
}
\examples{
df <- sim_lead_data(n = 1000, dmin = as.Date("2020-01-01"))

}
