% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipper_seq.R
\name{skipper_seq}
\alias{skipper_seq}
\title{Create a sequence that avoids some numbers}
\usage{
skipper_seq(start, length, skip)
}
\arguments{
\item{start}{An integer: the first number in the sequence.}

\item{length}{An integer: the length of the sequence.}

\item{skip}{A vector of integers that should not be included in the sequence.}
}
\value{
A numeric vector of length \code{length}.
}
\description{
\code{skipper_seq} generates a numeric sequence skipping over any integers provided in \code{skip}.
}
\examples{
s <- skipper_seq(start = 1000, length = 50, skip = c(1000, 1004:1008, 1045:1055))
}
