% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-street_names.R
\docType{data}
\name{street_names}
\alias{street_names}
\title{Street names}
\format{
A character vector.
}
\source{
Taken from 1) lists of street names in Kansas City, Lawrence, Manhattan, Topeka, and Wichita, KS at geographic.org (https://geographic.org/streetview/usa/ks/index.html), and 2) a Washington Post article on the most popular street names in the United States (https://www.washingtonpost.com/blogs/govbeat/wp/2015/03/06/these-are-the-most-popular-street-names-in-every-state/).
}
\usage{
street_names
}
\description{
A vector of street names.
}
\keyword{datasets}
