% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-street_sfx.R
\docType{data}
\name{street_sfx}
\alias{street_sfx}
\title{Street suffixes}
\format{
A dataframe.
}
\source{
Created for use in address cleaning and address creation functions.
}
\usage{
street_sfx
}
\description{
A table of street suffixes, with both full and abbreviated forms.
}
\keyword{datasets}
