% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_date_range.R
\name{subset_date_range}
\alias{subset_date_range}
\title{Subset a dataframe by a date variable}
\usage{
subset_date_range(df, var, range)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{A variable name from \code{df} containing dates.}

\item{range}{A character vector of length 1-2.}
}
\value{
A dataframe.
}
\description{
Subset a dataframe by year(s), year-quarter(s), or dates. Supply the date variable \code{var} and \code{range}, a character vector of length 1-2 containing
\itemize{
\item A single year (e.g., "2023") or year-quarter (e.g., "2023q1")
\item A year and/or year-quarter pair (start & end)
\item A pair (start & end) of dates (e.g., c("2022-02-02", "2023-03-03"))
}
}
\examples{
df <- data.frame(date = sample(seq.Date(as.Date("2020-01-01"), as.Date("2022-12-31"),
                                        by = "day"), size = 500))
df_new <- subset_date_range(df, var = "date", range = "2021q3")
}
