% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dates.R
\name{subset_dates}
\alias{subset_dates}
\title{Subset a dataframe by a date variable}
\usage{
subset_dates(df, var, range1, range2 = NULL, silent = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{The date variable by which to subset \code{df}.}

\item{range1, range2}{A string representing a range as a year, year-quarter, or date (YYYY-MM-dd).}

\item{silent}{Logical: silence message output if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
Subset a dataframe by one or two or strings representing years, year-quarters, or full dates.
}
\details{
Accepted values for \code{range1} and \code{range2} include 4-character years (e.g., "2019"), 6-character year-quarters (e.g., "2019q4"), or 10-character dates (e.g., "2019-10-25"). If \code{range2 == NULL} (the default), \code{df} will be subset by the full date range represented by \code{range1}. If \code{range2 != NULL}, \code{df} will be subset from the earliest date represented by \code{range1} to the latest date represented by \code{range2}. Note that if both \code{range1} and \code{range2} are provided, they must be either 1) both years or year-quarters or 2) both dates.
}
\examples{
dates <- seq.Date(as.Date("2020-01-01"), as.Date("2022-12-31"), by = "day")

df <- data.frame(
  date = dates,
  number = as.numeric(dates),
  character = as.character(dates)
)

dfa <- subset_dates(df, "date", "2021")
dfb <- subset_dates(df, "date", "2022q3")
dfc <- subset_dates(df, "date", "2021q4", "2022q1")
dfx <- subset_dates(df, "date", "2022-01-17", "2022-10-25")

}
