% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-test_addresses.R
\docType{data}
\name{test_addresses}
\alias{test_addresses}
\title{Places and addresses}
\format{
A dataframe.
}
\source{
Created for testing functions that manipulate, clean, and validate addresses.
}
\usage{
test_addresses
}
\description{
A table of places and addresses for testing functions.
}
\keyword{datasets}
