% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{test_results}
\alias{test_results}
\alias{check_results}
\alias{clean_results}
\alias{parse_results}
\title{Check, clean, and parse test results}
\usage{
check_results(results)

clean_results(results)

parse_results(results)
}
\arguments{
\item{results}{A vector of test results.}
}
\value{
A vector the same length as \code{results}.
\itemize{
\item \code{check_results()} and \code{clean_results()} return a vector the same length as \code{results}.
\item \code{parse_results()} returns a list containing two vectors, \code{sign} and \code{number}, the same length as \code{results}.
}
}
\description{
\itemize{
\item \code{check_results()} finds unexpected test result values and returns a logical vector where \code{TRUE} indicates a valid result.
\item \code{clean_results()} cleans common errors and converts "Not/None detected" to \verb{< 1.0}.
\item \code{parse_results()} separates the greater- or less-than sign (\code{>} or \code{<}) from the number in the result.
}
}
\concept{lab result processing functions}
