% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_followups.R
\name{track_followups}
\alias{track_followups}
\title{Track follow-ups for lead test records}
\usage{
track_followups(df, silent = FALSE)
}
\arguments{
\item{df}{A dataframe of lead test records.}

\item{silent}{Logical: silence progress indicator if \code{TRUE}.}
}
\value{
A dataframe with the new variable \code{days_to_followup}.
}
\description{
This function gets the number of days between the lab collection date and the date of the follow-up for tests with elevated results.
}
