% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-unit_pfx.R
\docType{data}
\name{unit_pfx}
\alias{unit_pfx}
\title{Unit prefixes}
\format{
A dataframe.
}
\source{
Created for use in address cleaning and address simulation.
}
\usage{
unit_pfx
}
\description{
A table of unit prefixes, with both full and abbreviated forms.
}
\keyword{datasets}
