% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps_lookup.R
\name{usps_lookup}
\alias{usps_lookup}
\title{Validate addresses at USPS.com}
\usage{
usps_lookup(
  df,
  row_id = NULL,
  var = c("street", "unit", "city", "state", "zip")
)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{var}{A vector of variable names for the following address components in this order: 1) house number and street, 2) unit number, 3) city, 4) state, and 5) zip code. Defaults to \code{c("street", "unit", "city", "state", "zip")}.}
}
\value{
A dataframe of results from USPS.com.
}
\description{
This function submits a table of addresses to the ZIP Code by Address lookup tool at USPS.com for validation.
}
\details{
The \code{row_id} argument is optional but recommended to easily match the results to the submitted addresses. In the returned dataframe, \code{n_row_src} corresponds to the row number of the input dataframe, and \code{n_result} indicates if multiple addresses were returned for any submitted address.

Resources
\itemize{
\item \href{https://tools.usps.com/zip-code-lookup.htm?byaddress}{ZIP Code by Address} lookup tool
\item \href{https://www.accuzip.com/webhelp/Appendix/DPV_Codes_and_Information.htm}{Information about Delivery Point Validation (DPV)}
}
}
\examples{
df <- data.frame(
  id = c("KDHE Division of Public Health", "Curtis State Office Bldg"),
  street = c("1000 SW Jackson", "1000 SW Jackson"),
  unit = c("Suite 540", NA),
  city = c("Topeka", "Topeka"),
  state = c("KS", "KS"),
  zip = c("66612", "66612")
)

\dontrun{
df_results <- usps_lookup(df, row_id = "id")
}

}
