% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_check.R
\name{var_check}
\alias{var_check}
\title{Check variable names in a dataframe}
\usage{
var_check(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{A character vector.}
}
\value{
Nothing is returned if all elements of \code{var} are present in \code{df}.
}
\description{
This is used mainly inside other functions to check for the presence of necessary variables. If any element of \code{var} is not present in \code{df}, \code{var_check()} stops function execution and returns an error listing the missing variables.
}
\examples{
x <- c("height", "age")
var_check(Loblolly, var = x)

# Strings are case-sensitive
try(var_check(Loblolly, "seed"))

}
