% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_md_url.R
\name{build_md_url}
\alias{build_md_url}
\title{Build a Melissa Data Personator URL request}
\usage{
build_md_url(df, row_id)
}
\arguments{
\item{df}{A dataframe of addresses including the \code{street_geo} variable.}

\item{row_id}{A unique row identifier. Should be \code{address_id}.}
}
\value{
A vector of URLs.
}
\description{
Build a URL to query the Melissa Data Personator Consumer Web Service API. The URL is a request for a response from the Personator service in a JSON format.
}
\section{Address processing workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item compare...
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=submit_to_md]{submit_to_md()}}
}
}

\section{Address validation and geocoding}{

The address processing workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for the request are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{clean_values}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{submit_to_md}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
