% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_md_url.R
\name{build_md_url}
\alias{build_md_url}
\title{Build a Melissa Data Personator request URL}
\usage{
build_md_url(
  df,
  row_id = "md_id",
  street = "street",
  unit = NULL,
  city = "city",
  state = "state",
  zip = "zip"
)
}
\arguments{
\item{df}{A dataframe of cleaned addresses.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{street}{The name of the variable containing the street address. Unit information should be included if the \code{unit} argument is left \code{NULL}.}

\item{unit}{The name of the variable containing the unit component of the street address (if the unit is not already included in \code{street}).}

\item{city}{The name of the variable containing the city.}

\item{state}{The name of the variable containing the state.}

\item{zip}{The name of the variable containing the zip code.}
}
\value{
A vector of URLs.
}
\description{
Build a URL to query the Melissa Data Personator Consumer Web Service API. The URL is a request for a response from the Personator service in JSON format.
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{clean_address}()},
\code{\link{clean_street_address}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
