% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbls_lab_table.R
\name{cbls_lab_table}
\alias{cbls_lab_table}
\title{Create a CBLS Lab Results table}
\usage{
cbls_lab_table(df, key, row_id, ref_labs)
}
\arguments{
\item{df}{A dataframe of records prepared for CBLS submission.}

\item{key}{A dataframe returned by \code{\link[=cbls_table_key]{cbls_table_key()}}.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{ref_labs}{A reference list of classified lab names.}
}
\value{
A dataframe formatted as a Lab Results table per CBLS guidelines.
}
\description{
Create a CBLS Lab Results table
}
