% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_seq.R
\name{check_date_seq}
\alias{check_date_seq}
\title{Check date sequences}
\usage{
check_date_seq(...)
}
\arguments{
\item{...}{Vectors containing dates.}
}
\value{
A logical vector of the same length as the arguments.
}
\description{
This function checks whether or not a sequence of dates is in order from earliest to latest. Any number of date vectors can be checked. The function expects dates to progress through time in argument order from left to right. Elements at the same index will be compared, as in \code{date1[1] < date2[1]}, \code{date2[1] < date3[1]}, etc. Equivalent dates are evaluated as being in order. \code{NA}s are omitted during evaluation.
}
\examples{
dates <- seq.Date(
  as.Date("2023-01-01", "\%Y-\%m-\%d"),
  as.Date("2023-12-31", "\%Y-\%m-\%d"),
  by = "day"
)
dates <- c(dates, rep(NA, times = 10))

n_rows <- 500

df <- data.frame(
  d1 = sample(dates, n_rows, replace = TRUE),
  d2 = sample(dates, n_rows, replace = TRUE),
  d3 = sample(dates, n_rows, replace = TRUE),
  d4 = sample(dates, n_rows, replace = TRUE)
)

df$ordered <- check_date_seq(df$d1, df$d2, df$d3, df$d4)

}
