% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_seq.R
\name{check_date_seq}
\alias{check_date_seq}
\title{Check date sequences in a dataframe}
\usage{
check_date_seq(df, dates, row_id = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{dates}{A character vector containing the names of date variables in \code{df} listed in the expected sequence, i.e., early to late.}

\item{row_id}{A variable name for a unique row identifier (optional).}
}
\value{
If any dates are out of sequence, a dataframe containing the affected rows is returned. Only the \code{row_id} and \code{dates} variables are returned from \code{df}.
}
\description{
This function checks the sequence of any dates added to the \code{dates} argument. Any number of date variables can be checked. Equivalent dates are evaluated as being in sequence.
}
\examples{
rows <- 100

dfA <- data.frame(
  id = 1:rows,
  d1 = sample(
    seq.Date(
      from = as.Date("2020-01-01"),
      to = as.Date("2022-12-31"),
      by = "day"
    ),
    size = rows
  ),
  d2 = sample(
    seq.Date(
      from = as.Date("2020-01-01"),
      to = as.Date("2022-12-31"),
      by = "day"
    ),
    size = rows
  )
)

}
