% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean dirty addresses}
\usage{
clean_address(df, type, var = "street", row_id = "address_id")
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{type}{The type of cleaning to perform. One of \code{c("digit_char", "dir_street", "embed_punct", "house_frac", "nondigit", "nonres", "nth", "num_dir", "num_only", "num_word", "numsign", "phone", "pobox", "rep_seg", "symbol", "unit", "unknown")}.}

\item{var}{A variable in \code{df} containing the targeted address component.}

\item{row_id}{A unique row identifier variable in \code{df}. Defaults to \code{address_id}.}
}
\value{
A dataframe containing only the rows that have been cleaned.
}
\description{
This function applies a pattern to clean street address strings in \code{df$var} based on the \code{type} chosen. The returned dataframe has two new columns, \code{removed_text} and \code{replacement_text}. Only cleaned rows are returned. After visually confirming the results and making any needed adjustments to the returned dataframe, use \code{\link[=replace_values]{replace_values()}} to substitute the value in \code{replacement_text} for the original value in \code{var}.
}
\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
