% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_values.R
\name{clean_values}
\alias{clean_values}
\title{Clean addresses in a dataframe}
\usage{
clean_values(df, var, id_var, type)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} containing street addresses.}

\item{id_var}{A variable name in \code{df} that serves as a unique row identifier.}

\item{type}{The type of cleaning to perform.}
}
\value{
A dataframe containing only the rows that have been cleaned.
}
\description{
This function applies a pattern to clean street address strings in \code{df$var} based on the \code{type} chosen. The returned dataframe has two new columns, \code{removed_text} and \code{replacement_text}. Only cleaned rows are returned. After visually confirming the results and making any needed adjustments to the returned dataframe, use \code{\link[=replace_values]{replace_values()}} to substitute the value in \code{replacement_text} for the original value in \code{var}.
}
\seealso{
Other address processing functions: 
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
