% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_epitrax.R
\name{config_epitrax}
\alias{config_epitrax}
\title{Configure the EpiTrax source file}
\usage{
config_epitrax(
  df,
  var_dates = etmstuff::epitrax_vars$date,
  var_final = etmstuff::epitrax_vars$final
)
}
\arguments{
\item{df}{A dataframe.}

\item{var_dates}{The EpiTrax date variables to format.}

\item{var_final}{The order for EpiTrax variables in the returned dataframe.}
}
\value{
A list containing two dataframes:
\itemize{
\item \code{data} contains all original variables, \code{row_id_src}, and \code{record_id_src}.
\item \code{keys} contains \code{row_id_src}, \code{patient_id}, \code{patient_record_number}, and \code{record_id_src}.
}
}
\description{
This function configures the EpiTrax source file in preparation for data processing. It does the following:
}
\details{
\itemize{
\item Adds a unique identifier (\code{record_id_src}) based on the values in the unmodified source file.
\item Adds a unique row identifier (\code{row_id_src}).
\item Formats all date variables as "YYYY-MM-DD".
\item Reorders the columns according to \code{etmstuff::epitrax_vars$all}.
}
}
