% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_ksde_project.R
\name{document_ksde_project}
\alias{document_ksde_project}
\title{Document a KS Tracking Data Explorer project}
\usage{
document_ksde_project(
  proj_dir,
  type = c("display", "geo", "mw", "version"),
  overwrite = FALSE,
  test_drive = FALSE
)
}
\arguments{
\item{proj_dir}{The name of the RStudio project to document.}

\item{type}{The project type.
\itemize{
\item \code{"display"} = display data
\item \code{"geo"} = core geography tables
\item \code{"mw"} = message widget table
\item \code{"version"} = version table
}}

\item{overwrite}{If the destination directory for the project already exists, it will be removed when \code{overwrite} is set to \code{TRUE}.}

\item{test_drive}{Replace the destination path with the desktop ("C:/Users/eliot.monaco/OneDrive - State of Kansas, OITS/Desktop/") if \code{TRUE}.}
}
\value{
No value returned.
}
\description{
Document a KS Tracking Data Explorer project
}
