% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epitrax_lead_vars.R
\docType{data}
\name{epitrax_lead_vars}
\alias{epitrax_lead_vars}
\title{EpiTrax variables used in lead data processing}
\format{
A list of three vectors.
}
\source{
Created for lead data processing.
}
\usage{
epitrax_lead_vars
}
\description{
Variable names contained in the EpiTrax lead data export file. Used in \code{\link[=config_epitrax]{config_epitrax()}}. The list contains three vectors:
\itemize{
\item \code{export}: All variable names in the EpiTrax export file
\item \code{ordered}: The order of variables for data processing
\item \code{date}: EpiTrax variables containing date values
}
}
\keyword{datasets}
