% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epitrax_vars.R
\docType{data}
\name{epitrax_vars}
\alias{epitrax_vars}
\title{EpiTrax variables}
\format{
A list.
}
\source{
EpiTrax lead data dictionary.
}
\usage{
epitrax_vars
}
\description{
A list of EpiTrax variables used in lead record data processing. The list contains the following vectors:
\itemize{
\item \code{all}: All exported variables
\item \code{final}: Variables kept in the final data set
}
}
\keyword{datasets}
