% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_close_tests.R
\name{find_close_tests}
\alias{find_close_tests}
\title{Find similar tests on different days}
\usage{
find_close_tests(df, days, silent = FALSE)
}
\arguments{
\item{df}{A dataframe of cleaned, deduplicated lead test records.}

\item{days}{The number of days to look ahead for similar tests.}

\item{silent}{Logical: silence progress indicator if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
This function finds tests that fit the criteria for duplicates (the same person, test result, and specimen source) but occur on different days.
}
