% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followup_overdue_flag.R
\name{followup_overdue_flag}
\alias{followup_overdue_flag}
\title{Flag lead tests for overdue follow-up}
\usage{
followup_overdue_flag(df)
}
\arguments{
\item{df}{A dataframe of lead test records.}
}
\value{
A dataframe with the new variable \code{followup_overdue}.
}
\description{
This function determines whether the follow-up for each test is overdue according to the KDHE Elevated Blood Lead Investigation Guideline document. The function is used in the lead dashboard.
}
