% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get_date_element}
\alias{get_date_element}
\alias{get_year}
\alias{get_quarter}
\title{Get year(s) or quarter(s) from a vector of dates}
\usage{
get_year(dates)

get_quarter(dates)
}
\arguments{
\item{dates}{A vector of dates, or a vector of strings that can be formatted as dates.}
}
\value{
\itemize{
\item \code{get_year()} returns a numeric vector of years.
\item \code{get_quarter()} returns a numeric vector of digits \code{1:4}.
}
}
\description{
\itemize{
\item \code{get_year()} returns the unique year(s) represented in a vector of dates.
\item \code{get_quarter()} returns the unique quarter(s) represented in a vector of dates.
}
}
\examples{
d <- seq.Date(from = as.Date("1980-10-01"), to = as.Date("1981-02-28"), by = "week")
get_year(d)
get_quarter(d)
}
