% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_distinct_rows.R
\name{id_distinct_rows}
\alias{id_distinct_rows}
\title{Assign a sequential ID number to distinct rows}
\usage{
id_distinct_rows(
  df,
  id_name,
  prefix = NULL,
  var = colnames(df),
  seq_start = 1,
  digits = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{id_name}{A name for the new ID variable.}

\item{prefix}{An optional prefix of characters to add to all ID values.}

\item{var}{The variable names whose unique values the new ID will be based on. Defaults to \code{colnames(df)}.}

\item{seq_start}{An integer for the beginning of the ID sequence. Defaults to \code{1}.}

\item{digits}{An integer for the number of digits in the ID (including leading zeros). Defaults to \code{nchar(nrow(df) + seq_start)}.}
}
\value{
A dataframe.
}
\description{
This function assigns a sequential ID number to distinct rows in a dataframe  based on selected variables (\code{var}). The new ID is a string, with the number of characters equal to \code{nchar(nrow(df))} (if there are 100 rows, the new ID will have three characters).
}
\examples{
n_rows <- 50

df <- data.frame(
  x = sample(c("cat", "horse", "howler monkey"), size = n_rows, replace = TRUE),
  y = sample(month.name, size = n_rows, replace = TRUE),
  z = seq(1, n_rows)
)

df <- id_distinct_rows(df, id_name = "new_id", prefix = "A", var = c("x", "y"))

}
