% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_results.R
\name{lab_results}
\alias{lab_results}
\alias{check_lab_results}
\alias{clean_lab_results}
\alias{parse_lab_results}
\title{Check, clean, and parse lab results}
\usage{
check_lab_results(df, var)

clean_lab_results(df, var)

parse_lab_results(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{The variable name from \code{df} with lab results to check, clean, or parse.}
}
\value{
A dataframe.
}
\description{
\itemize{
\item \code{check_lab_results()} finds unexpected values in the lab results and returns a dataframe with only those records.
\item \code{clean_lab_results()} cleans common errors in the results and converts "Not/None detected" to \code{0}. The cleaned results appear in a new variable, \code{lab_results_cleaned}.
\item \code{parse_lab_results()} creates two new variables: \code{lab_results_symbol}, containing the comparison operator (\code{<} or \code{>}), and \code{lab_results_number}, containing the numeric result.
}
}
\concept{lab result processing functions}
