% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_results.R
\name{lab_results}
\alias{lab_results}
\alias{check_lab_results}
\alias{clean_lab_results}
\alias{parse_lab_results}
\title{Check, clean, and parse lab results}
\usage{
check_lab_results(df, var)

clean_lab_results(df, var)

parse_lab_results(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{A variable name from \code{df} containing lab results.}
}
\value{
A dataframe.
}
\description{
\itemize{
\item \code{check_lab_results()}
\item \code{clean_lab_results()}
\item \code{parse_lab_results()}
}
}
