% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_results_table.R
\name{md_results_table}
\alias{md_results_table}
\title{Count result codes in the Melissa Data response}
\usage{
md_results_table(df, var = "Results")
}
\arguments{
\item{df}{A dataframe of Melissa Data responses returned by \code{\link[=md_batch_request]{md_batch_request()}}.}

\item{var}{The variable containing the result codes. The default is \code{"Results"}.}
}
\value{
A dataframe.
}
\description{
This function creates a table with counts of the result codes present in \code{df$Results}.
}
\section{Address validation workflow}{

\enumerate{
\item Prepare addresses
\itemize{
\item \code{\link[=pull_addresses]{pull_addresses()}}
}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_address]{clean_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item Validate/geocode addresses
\itemize{
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
\item Review results
\itemize{
\item \code{\link[=md_results_table]{md_results_table()}}
}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

