% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_results_table.R
\name{md_results_table}
\alias{md_results_table}
\title{Tally addresses sent to Melissa Data by result code}
\usage{
md_results_table(df, var = "Results")
}
\arguments{
\item{df}{A dataframe of Melissa Data responses.}

\item{var}{The variable containing result codes. The default is \code{Results}.}
}
\value{
A dataframe tallying addresses by certain Melissa Data result codes.
}
\description{
This function takes a dataframe of Melissa Data responses returned by \code{\link[=md_batch_request]{md_batch_request()}} and returns a table tallying the responses by result codes that are relevant for the lead address validation and geocoding process.
}
