% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_hist.R
\name{my_hist}
\alias{my_hist}
\title{Plot a histogram with \code{plotly}}
\usage{
my_hist(x, bin_size = 1, x_label = NULL, title = NULL, bar_gap = NULL)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{bin_size}{The size of each bin.}

\item{x_label}{The x-axis label.}

\item{title}{The plot title.}

\item{bar_gap}{A number from 0 to 1. The gap between bars as a fraction of the plot.}
}
\value{
None.
}
\description{
Plot a histogram with \code{plotly}
}
\examples{
x <- rnorm(100)

\dontrun{
my_hist(x)
}

}
