% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_kable.R
\name{my_kable}
\alias{my_kable}
\title{Create a table using \code{kableExtra} formatting}
\usage{
my_kable(
  df,
  caption = NULL,
  bootstrap_opts = NULL,
  row_names = FALSE,
  format = "html"
)
}
\arguments{
\item{df}{A dataframe.}

\item{caption}{A table caption.}

\item{bootstrap_opts}{A character vector of bootstrap options to pass to \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}. Options are \code{basic}, \code{striped}, \code{bordered}, \code{hover}, \code{condensed}, \code{responsive}, and \code{none},}

\item{row_names}{Logical: row names will be \code{rownames(df)} if \code{TRUE}.}

\item{format}{A value passed to the \code{format} argument in \code{\link[kableExtra:kbl]{kableExtra::kbl()}} (see that function's documentation for options). Default is \code{"html"}.}
}
\value{
None.
}
\description{
This function returns a table formatted using the \link{kableExtra} package.
}
\examples{
df <- mtcars[1:10,]

tbl <- my_kable(df,
  caption = "Check out my sweeeeeet table!",
  row_names = TRUE,
  bootstrap_opts = "bordered"
)

\dontrun{
# Apply conditional formatting to text in `mtcars$hp`
tbl \%>\%
  kableExtra::column_spec(
    column = which(colnames(df) == "hp") + 1,
    color = ifelse(df$hp > 100, "#FF801E", "#EFECE4")
  )
}

}
