% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_kable.R
\name{my_kable}
\alias{my_kable}
\title{Create custom table using \code{kable()}}
\usage{
my_kable(
  df,
  row_names = FALSE,
  alt_row_bg = FALSE,
  caption = NULL,
  bootstrap_opts = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{row_names}{Logical: use \code{rownames(df)} if \code{TRUE}.}

\item{alt_row_bg}{Logical: alternate row background color if \code{TRUE}.}

\item{caption}{A table caption (optional).}

\item{bootstrap_opts}{Bootstrap options to pass to \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.}
}
\value{
A table formatted using functions in the \link{kableExtra} package.
}
\description{
This function returns a table formatted using functions in the \link{kableExtra} package. Additional formatting can be added to the returned table, as in the example. The following colors are used:
\itemize{
\item Normal text: "#EFECE4"
\item Emphasized text: "#FF801E"
\item Header row background: "#0B1433"
\item Body row background: "#434E77"
\item Alternate row background: "#1D274F"
}
}
\examples{
df <- mtcars[1:10,]

tbl <- my_kable(
  df,
  row_names = TRUE,
  alt_row_bg = TRUE,
  caption = "Check out my sweeeeeet table!",
  bootstrap_opts = "bordered"
)

\dontrun{
# Apply conditional formatting to text in `mtcars$hp`
tbl \%>\%
  kableExtra::column_spec(
    column = which(colnames(df) == "hp") + 1,
    color = ifelse(df$hp > 100, "#FF801E", "#EFECE4")
  )
}

}
