% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xl_table.R
\name{my_xl_table}
\alias{my_xl_table}
\title{Save a dataframe to an Excel workbook with custom formatting}
\usage{
my_xl_table(
  df,
  file,
  style_cols = NULL,
  as_table = TRUE,
  sheet = "Sheet1",
  overwrite = TRUE
)
}
\arguments{
\item{df}{A dataframe.}

\item{file}{A file path. \code{".xlsx"} will be appended if not present.}

\item{style_cols}{A vector of variable names in \code{df} to highlight.}

\item{as_table}{Logical: format data as a table if \code{TRUE}.}

\item{sheet}{A name for the worksheet.}

\item{overwrite}{Logical: overwrite existing file if \code{TRUE}.}
}
\value{
An Excel workbook.
}
\description{
This function produces an Excel workbook with the dataframe \code{df} formatted as a table. Variable names passed to \code{style_cols} will be highlighted in red. If \code{nrow(df) == 0}, \code{df} will be returned invisibly without producing an Excel workbook.
}
\details{
\code{my_xl_table()} is built on functions from the \link{openxlsx} package.
}
\examples{
\dontrun{
my_xl_table(mtcars, file = "mtcars_workbook", style_cols = c("mpg", "cyl"))
}

}
