% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_xlsx.R
\name{my_xlsx}
\alias{my_xlsx}
\title{Save XLSX file with custom settings using \code{write.xlsx()}}
\usage{
my_xlsx(x, file)
}
\arguments{
\item{x}{A dataframe or list of dataframes.}

\item{file}{A file path.}
}
\value{
An XLSX file.
}
\description{
Save an object (a dataframe or list) as an XLSX file with custom table formatting using \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}. If \code{x} is a list of dataframes, each dataframe will be saved as a separate sheet.
}
\examples{
\dontrun{
my_xlsx(mtcars, "mtcars.xlsx")
}

}
