% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-names.R
\docType{data}
\name{names}
\alias{names}
\title{First and last names}
\format{
A list.
}
\source{
First names obtained from the Social Security Administration website: \href{https://www.ssa.gov/oact/babynames/decades/century.html}{Top Names Over the Last 100 Years}. Last names downloaded from the US Census website: \href{https://www.census.gov/topics/population/genealogy/data/2010_surnames.html}{Frequently Occurring Surnames from the 2010 Census}.
}
\usage{
names
}
\description{
A list of first and last names to use in simulated data. The list contains the following vectors:
\itemize{
\item \code{first_male}: 100 male first names
\item \code{first_female}: 100 female first names
\item \code{last}: 1000 last names
}
}
\keyword{datasets}
