% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_street_addresses.R
\name{parse_street_addresses}
\alias{parse_street_addresses}
\title{Parse street addresses using \code{postmastr} workflow}
\usage{
parse_street_addresses(df, street = "street")
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{street}{The street address variable name.}
}
\value{
A dataframe.
}
\description{
Parse street addresses using functions from the \code{postmastr} package. \code{df} requires an \code{address_id} variable before using \code{parse_street_addresses()}. The output contains a new variable, \code{street_unit}, which combines both \code{street} and \code{unit}.

\code{postmastr} is available in a development version from GitHub.

\if{html}{\out{<div class="sourceCode">}}\preformatted{remotes::install_github("slu-openGIS/postmastr")
}\if{html}{\out{</div>}}
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item \code{\link[=compare_parsed_street]{compare_parsed_street()}}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{send_md_request}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
