% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_max_test_result.R
\name{pick_max_test_result}
\alias{pick_max_test_result}
\title{Select the test with the highest result for each person}
\usage{
pick_max_test_result(
  df,
  eval = c("all_tests", "cfmd_preferred", "cfmd_only"),
  silent = FALSE
)
}
\arguments{
\item{df}{A dataframe.}

\item{eval}{\itemize{
\item \code{all_tests}: pick the highest result among all tests for each person.
\item \code{cfmd_preferred}: if both confirmed and unconfirmed tests are present for a person, pick the highest confirmed test result; otherwise pick the highest unconfirmed test result.
\item \code{cfmd_only}: if both confirmed and unconfirmed tests are present for a person, pick the highest confirmed test result and ignore any unconfirmed test results.
}}

\item{silent}{Logical: silence progress indicator if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
This function returns a maximum of one test per person from \code{df}. Confirmed and unconfirmed tests are handled differently based on the mode selected in the \code{eval} argument. Tests are filtered first by the \code{patient_id} variable. Test results are evaluated by the value in \code{lab_result_number} only; \code{lab_result_symbol} is not considered.
}
\details{
The \code{test_reason} variable is created by the function \code{\link[=assign_test_reason]{assign_test_reason()}}. See that function's documentation for definitions of \code{test_reason} values.

Confirmed results are from tests with these \code{test_reason} values:
\itemize{
\item \code{ven_cfm_init}
\item \code{cap_cfm_elev}
\item \code{cap_cfm_nonelev}
\item \code{ven_cfm_elev}
\item \code{ven_cfm_nonelev}
\item \code{ven_flw}
}

Unconfirmed results are from tests with these \code{test_reason} values:
\itemize{
\item \code{cap_scrn}
\item \code{unknown}
}
}
