% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_addresses.R
\name{pull_addresses}
\alias{pull_addresses}
\title{Pull addresses from EpiTrax data}
\usage{
pull_addresses(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe of addresses only.
}
\description{
Pull addresses from EpiTrax data
}
\seealso{
Other address processing functions: 
\code{\link{clean_values}()},
\code{\link{replace_values}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
