% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_addresses.R
\name{pull_addresses}
\alias{pull_addresses}
\title{Pull addresses from EpiTrax data}
\usage{
pull_addresses(
  df,
  row_id,
  var = c("coll_add_street", "coll_add_unit", "coll_add_city", "coll_add_state",
    "coll_add_zip", "coll_add_county")
)
}
\arguments{
\item{df}{A dataframe of records from EpiTrax.}

\item{row_id}{A unique row identifier variable in \code{df}.}

\item{var}{A vector of variable names for the following address components in this order: 1) house number and street, 2) unit number, 3) city, 4) state, 5) zip code, and 6) county. Defaults to \code{c("coll_add_street", "coll_add_unit", "coll_add_city", "coll_add_state", "coll_add_zip", "coll_add_county")}.}
}
\value{
A dataframe of addresses. Columns are renamed to their generic address components. \code{street_src} is a copy of \code{street}, which preserves the original string for comparison during cleaning.
}
\description{
Subset the complete address at lab collection from EpiTrax data. Addresses with insufficient data for validation via Melissa Data are omitted.
}
\section{Address validation workflow}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service.
\enumerate{
\item Prepare addresses
\itemize{
\item \code{\link[=pull_addresses]{pull_addresses()}}
}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_address]{clean_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item Validate/geocode addresses
\itemize{
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=send_md_request]{send_md_request()}}
}
\item Review results
\itemize{
\item \code{\link[=md_results_table]{md_results_table()}}
}
}
}

\section{Melissa Data}{

To use the Check action to validate and correct addresses, Melissa Data requires, at minimum, one of the follow combinations of address fields:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

For more information on constructing the request, see the \href{https://www.melissa.com/quickstart-guides/personator-consumer}{Personator Consumer Quick Start Guide} and the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer:Request}{Personator Consumer Request wiki page}. Additional information can be found at the \href{https://wiki.melissadata.com/index.php?title=Personator_Consumer}{Melissa Data wiki}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_address}()},
\code{\link{md_results_table}()},
\code{\link{replace_values}()},
\code{\link{send_md_request}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
