% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_addresses.R
\name{pull_addresses}
\alias{pull_addresses}
\title{Pull addresses from EpiTrax data}
\usage{
pull_addresses(df)
}
\arguments{
\item{df}{A dataframe of records from EpiTrax.}
}
\value{
A dataframe of addresses and keys.
}
\description{
Subset address fields from EpiTrax data. The returned dataframe has one row per address. \code{recno} and the added variable \code{address_src} serve as a key to link each address to the original record. Addresses with insufficient data for geocoding are omitted.
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_values]{validate_values()}}
\item \code{\link[=clean_values]{clean_values()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item compare function...
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=submit_to_md]{submit_to_md()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_values}()},
\code{\link{parse_street_addresses}()},
\code{\link{replace_values}()},
\code{\link{submit_to_md}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
