% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_addresses.R
\name{pull_addresses}
\alias{pull_addresses}
\title{Pull addresses from EpiTrax data}
\usage{
pull_addresses(df, row_id)
}
\arguments{
\item{df}{A dataframe of records from EpiTrax.}

\item{row_id}{A unique row identifier variable in \code{df}.}
}
\value{
A dataframe of addresses. Columns are renamed to their generic address components. \code{street_src} is an additional street variable intended to preserve the street address from the source data for comparison to the cleaned and parsed street addresses (created later).
}
\description{
Subset the complete address at lab collection from EpiTrax data. Addresses with insufficient data for validation via Melissa Data are omitted.
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{md_request}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
