% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Replace dirty addresses with cleaned addresses}
\usage{
replace_values(
  df,
  var,
  df_src,
  var_src = "replacement_text",
  row_id = "address_id"
)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} whose values will be replaced.}

\item{df_src}{A dataframe with the variable supplied in \code{var_src} that contains the replacement text.}

\item{var_src}{A variable name in \code{df_src} that contains the replacement text. Defaults to \code{replacement_text.}}

\item{row_id}{A unique row identifier variable in \code{df}. Defaults to \code{address_id}.}
}
\value{
A dataframe.
}
\description{
Replace the dirty address values in \code{df$var} with the cleaned values in \code{df_src$var_src} generated by \code{\link[=clean_street_address]{clean_street_address()}}.
}
\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
