% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Replace dirty addresses with cleaned addresses}
\usage{
replace_values(df, var, row_id, source)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} whose values will be replaced.}

\item{row_id}{A variable name in \code{df} that serves as a unique row identifier.}

\item{source}{A dataframe with the variable \code{replacement_text}.}
}
\value{
A dataframe.
}
\description{
Replace the dirty address values in \code{df$var} with the cleaned values in \code{source$replacement_text} generated by \code{\link[=clean_street_address]{clean_street_address()}}.
}
\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{submit_to_md}()},
\code{\link{validate_values}()}
}
\concept{address processing functions}
