% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Replace dirty addresses with cleaned addresses}
\usage{
replace_values(
  df,
  var,
  df_src,
  var_src = "replacement_text",
  row_id = "address_id"
)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} whose values will be replaced.}

\item{df_src}{A dataframe with the variable supplied in \code{var_src} that contains the replacement text.}

\item{var_src}{A variable name in \code{df_src} that contains the replacement text. Defaults to \code{replacement_text.}}

\item{row_id}{A unique row identifier variable in \code{df}. Defaults to \code{address_id}.}
}
\value{
A dataframe.
}
\description{
Replace the dirty address values in \code{df$var} with the cleaned values in \code{df_src$var_src} generated by \code{\link[=clean_address]{clean_address()}}.
}
\section{Address validation workflow}{

\enumerate{
\item Prepare addresses
\itemize{
\item \code{\link[=pull_addresses]{pull_addresses()}}
}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_address]{clean_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item Validate/geocode addresses
\itemize{
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
\item Review results
\itemize{
\item \code{\link[=md_results_table]{md_results_table()}}
}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_address}()},
\code{\link{md_request}()},
\code{\link{pull_addresses}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
