% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_md_request.R
\name{send_md_request}
\alias{send_md_request}
\title{Submit requests to Melissa Data Personator}
\usage{
send_md_request(df, url = "md_url", url_open_mode = "")
}
\arguments{
\item{df}{A dataframe with the variable \code{md_url}.}

\item{url}{The name of the variable containing the URL requests for Melissa Data Personator.}

\item{url_open_mode}{See help for \code{\link[=url]{url()}} and Details section of the help for \code{\link[=download.file]{download.file()}} for an explanation of why \code{"wb"} may be necessary when running on Windows.}
}
\value{
A dataframe with responses from Melissa Data in additional columns.
}
\description{
This function submits the URL requests created by \code{\link[=build_md_url]{build_md_url()}} to the Melissa Data Personator Consumer Web Service API. The response is added to \code{df} as additional columns.
}
\section{Address validation workflow}{

\enumerate{
\item \code{\link[=pull_addresses]{pull_addresses()}}
\item Clean addresses
\itemize{
\item \code{\link[=validate_address]{validate_address()}}
\item \code{\link[=clean_street_address]{clean_street_address()}}
\item \code{\link[=replace_values]{replace_values()}}
}
\item \code{\link[=parse_street_addresses]{parse_street_addresses()}}
\item \code{\link[=compare_parsed_street]{compare_parsed_street()}}
\item \code{\link[=build_md_url]{build_md_url()}}
\item \code{\link[=md_batch_request]{md_batch_request()}}
}
}

\section{Melissa Data}{

The address validation workflow prepares addresses for validation by the Melissa Data Personator Consumer Web Service. Minimum data requirements for requests are either of these value combinations:
\itemize{
\item \code{street}, \code{city}, and \code{state}
\item \code{street} and \code{zip}
}

More information is at \url{https://wiki.melissadata.com/index.php?title=Personator_Consumer}.
}

\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{md_request}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{validate_address}()}
}
\concept{address processing functions}
