% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_address.R
\name{sim_address}
\alias{sim_address}
\title{Simulate addresses}
\usage{
sim_address(nrow, dirty = FALSE)
}
\arguments{
\item{nrow}{An integer to set the number of rows in the output.}

\item{dirty}{Logical: creates dirty data in \code{street}, \code{city}, and \code{zip} when set to \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
Create a dataframe of simulated addresses. Street names are from \code{street_names}, which is based on a list of the most common street names in the US. City, zip code, and county combinations are from \code{ks_city_zip}. Other elements (e.g., house numbers, directions, street suffixes, and units) are generated pseudorandomly using \code{sample()}. By setting \code{dirty = TRUE}, the function will alter values in \code{street}, \code{city}, and \code{zip} for the purpose of testing \code{\link[=validate_address]{validate_address()}} and \code{\link[=clean_address]{clean_address()}}.
}
\examples{
df <- sim_address(nrow = 1000, dirty = TRUE)

}
