% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize strings in a dataframe}
\usage{
standardize(df, uppercase = TRUE, vars_ignore = NULL, silence = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{uppercase}{Logical: change text to uppercase if \code{TRUE} (default).}

\item{vars_ignore}{A character vector naming which variables to ignore.}

\item{silence}{Logical: silence progress bar if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
Remove excess spaces from strings, replace empty cells with \code{NA}, and convert strings to uppercase (optional).
}
\examples{
df <- data.frame(x = c("pear", " kumquat ", "fuji   apple", " "),
                 y = 1:4)
df_stdz <- standardize(df)
}
