% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize strings in a dataframe}
\usage{
standardize(
  df,
  vars = NULL,
  vars_ignore = NULL,
  uppercase = TRUE,
  silent = FALSE
)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{A character vector naming which variables to standardize.}

\item{vars_ignore}{A character vector naming which variables to ignore.}

\item{uppercase}{Logical: change text to uppercase if \code{TRUE} (default).}

\item{silent}{Logical: silence output to console if \code{TRUE}.}
}
\value{
A dataframe.
}
\description{
Remove excess spaces from strings, replace empty cells with \code{NA}, and convert strings to uppercase (optional). If neither \code{vars} nor \code{vars_ignore} are supplied, all character variables in \code{df} are standardized.
}
\examples{
df <- data.frame(
  x = c("pear", " kumquat ", "fuji   apple", " "),
  y = 1:4
)
df_stdz <- standardize(df)
}
