% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{test_results}
\alias{test_results}
\alias{check_results}
\alias{clean_results}
\alias{parse_results}
\alias{flag_results}
\title{Check, clean, and parse test results}
\usage{
check_results(results)

clean_results(results)

parse_results(results)

flag_results(sign, number)
}
\arguments{
\item{results}{A vector of test results.}

\item{sign}{The \code{sign} vector returned by \code{parse_results()}}

\item{number}{The \code{number} vector returned by \code{parse_results()}}
}
\value{
\itemize{
\item \code{check_results()} and \code{flag_results()} return a logical vector the same length as the argument(s).
\item \code{clean_results()} returns a character vector the same length as the argument.
\item \code{parse_results()} returns a list containing \code{sign} (a character vector) and \code{number} (a numeric vector), each the same length as the argument.
}
}
\description{
\itemize{
\item \code{check_results()} finds results that don't match the expected pattern of a greater- or less-than sign and a number.
\item \code{clean_results()} cleans common errors and converts "Not/None detected" to \code{"< 1.0"}.
\item \code{parse_results()} separates the greater- or less-than sign from the number in the cleaned results.
\item \code{flag_results()} finds parsed results that require review.
}
}
