% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps_lookup.R
\name{usps_lookup}
\alias{usps_lookup}
\title{Look up addresses at USPS.com}
\usage{
usps_lookup(df, row_id = NULL)
}
\arguments{
\item{df}{A dataframe of addresses with the variables \code{street}, \code{unit}, \code{city}, \code{state}, and \code{zip}.}

\item{row_id}{A unique row identifier variable in \code{df} (optional).}
}
\value{
A dataframe of results from USPS.com.
}
\description{
This function submits a table of addresses to the ZIP Code by Address lookup tool at USPS.com and returns the lookup results. The variables \code{street}, \code{unit}, \code{city}, \code{state}, and \code{zip} are required in \code{df}. \code{row_id} is optional but recommended. The variables \code{n_row_src} and \code{n_result} are added to enable matching the results to the submitted addresses.
}
\details{
Resources
\itemize{
\item \href{https://tools.usps.com/zip-code-lookup.htm?byaddress}{ZIP Code by Address}
\item \href{https://www.accuzip.com/webhelp/Appendix/DPV_Codes_and_Information.htm}{Information about Delivery Point Validation (DPV)}
}
}
\examples{
df <- data.frame(
  id = c("KDHE BEPHI", "Curtis State Office Bldg"),
  street = c("1000 SW Jackson", "1000 SW Jackson"),
  unit = c("Suite 110", NA),
  city = c("Topeka", "Topeka"),
  state = c("KS", "KS"),
  zip = c("66612", "66612")
)

df_results <- usps_lookup(df, row_id = "id")

}
