% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_address.R
\name{validate_address}
\alias{validate_address}
\title{Validate address components}
\usage{
validate_address(df, var, type = var, max_dist = 0.1)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} containing the address component to validate.}

\item{type}{The type of address component to validate. One of \code{c("city", "zip")}. Defaults to value in \code{var}.}

\item{max_dist}{A value passed to \code{max.distance} in \code{\link[=agrep]{agrep()}} that determines the "maximum distance allowed for a match" to the elements in the reference list.}
}
\value{
A dataframe with a column of possible matches to replace the values in \code{var}.
}
\description{
This function checks an address component in \code{df$var} against elements in a reference list, e.g., cities in Kansas. If an exact match isn't found, it uses \code{\link[=agrep]{agrep()}} to find an approximate match or matches. The returned dataframe has one new column, \code{replacement_text}. Only rows without an exact match are returned. After visually confirming the results and making any needed adjustments to the returned dataframe, use \code{\link[=replace_values]{replace_values()}} to substitute the value in \code{replacement_text} for the original value in \code{var}.
}
\seealso{
Other address processing functions: 
\code{\link{build_md_url}()},
\code{\link{clean_street_address}()},
\code{\link{compare_parsed_street}()},
\code{\link{md_request}()},
\code{\link{parse_street_addresses}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()},
\code{\link{send_md_request}()}
}
\concept{address processing functions}
