% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_values.R
\name{validate_values}
\alias{validate_values}
\title{Validate address components}
\usage{
validate_values(df, var, ref, max_dist = 0.1)
}
\arguments{
\item{df}{A dataframe of addresses.}

\item{var}{A variable name in \code{df} containing values to validate.}

\item{ref}{A reference file with a list of values to validate \code{var} against.}

\item{max_dist}{A value passed to \code{max.distance} in \code{agrep()} that determines the "maximum distance allowed for a match" to the pattern in \code{ref}.}
}
\value{
A dataframe with a column of possible matches to replace the values in \code{var}.
}
\description{
This function checks the address component in var against values in a reference file. If an exact match isn't found, it uses \code{\link[=agrep]{agrep()}} to find an approximate match or matches. Only rows without an exact match are returned.
}
\seealso{
Other address processing functions: 
\code{\link{clean_values}()},
\code{\link{pull_addresses}()},
\code{\link{replace_values}()}
}
\concept{address processing functions}
